requirements.txt
sensors.py==1.0
simpleaudio==1.0.0
